Tweak code format script to ignore symlinked .cpp / .hpp files
Emacs likes to add hidden symlinks with names like '.#myfile.cpp' that don't actually point to real files as a locking mechanism, which cause the format check/apply scripts to fail. This tweaks the format scripts to ignore files that are symlinks (adds a -type f predicate to the find commands), which gets around this issue.