Preprocess input variance into BN denominator for cudnn (#1885)
* Add explicit cuda kernel to calculate what cuDNN describes as the inverse
variance. In reality, the backward cudnn kernel for BN requires 1.0f / sqrt(variance + eps),
which is the batchnorm denominator for each channel (a numerically stable inverse stddev).

This introduces op annotations for batch norm backprop and updates the cudnn_emitter to support the insertion of this cuda kernel when required.