[CPUCodegen] Introduce CPURuntimeContextCG for standalone codegen generation.
This patch introduces CPURuntimeContextCG. This class is aimed at
removing the dependency between nGraph and the generated code in
codegen mode. It will be used to hold the runtime context in
codegen mode and it will be emitted in the generated code. For now,
CPURuntimeContextCG only contains TBB's graph and global context.
Follow-up patches will migrate more members in CPURuntimeContext to
CPURuntimeContextCG for codegen mode.

Testing results:
  - Before: NGRAPH_CODEGEN=1 test/unit-test
    [----------] Global test environment tear-down
    [==========] 2503 tests from 54 test cases ran. (290406 ms total)
    [  PASSED  ] 2490 tests.

  - After: NGRAPH_CODEGEN=1 test/unit-test
    [----------] Global test environment tear-down
    [==========] 2503 tests from 54 test cases ran. (412616 ms total)
    [  PASSED  ] 2490 tests.