ReshapeElimination -> PlaidML; fixup ImplicitBcast
The PrefixReshapeElimination pass doesn't work correctly if the network's operation's input and output shape
compatibility is being validated; an implicit broadcast operation is required to get the shapes right.  This
change moves the implementation back to PlaidML (since it's seeming less generally useful, and PlaidML has an
implicit broadcast operation), and fixes both it and the implicit broadcast pass to only apply when the
downstream operation is an elementwise operation (since PlaidML doesn't support automatic broadcasting for
contractions).