[MLIR] Add config files for LIT
LIT is the LLVM testing tool that is used in LLVM and MLIR. It's
necessary to enable all the MLIR testing in nGraph. This will provide
a complementary finer grain level of testing than the one currently done
with gtest. LIT will allow testing each MLIR-based nGraph transformation
without having to run the whole nGraph pipeline. For example, we will be
able to write an nGraph dialect input, invoke an single MLIR transformation
on it and check that the output is the expected.

This PR adds the basic configuration files to enable LIT tool
in nGraph for MLIR Compiler. `affine_lowering/lit_test.mlir` is a
dummy test to exercise the new testing feature. Lit testing will be
enabled in CMAKE in a subsequent PR.